// types/booking.types.ts
/**
 * Booking API Types
 */

export interface BookingDates {
    checkin: string;
    checkout: string;
}

export interface BookingPayload {
    firstname: string;
    lastname: string;
    totalprice: number;
    depositpaid: boolean;
    bookingdates: BookingDates;
    additionalneeds?: string;
}

export interface BookingResponse {
    bookingid: number;
    booking: BookingPayload;
}

export interface BookingPostResponse {
    booking: BookingPayload;
    bookingid: number;
}